Add TS.MREVRANGE and extract Multi
