+++ lib/redis-time-series.rb
@@ -7,6 +7,7 @@ require 'redis/time_series/errors'
 require 'redis/time_series/aggregation'
 require 'redis/time_series/duplicate_policy'
 require 'redis/time_series/filters'
+require 'redis/time_series/multi'
 require 'redis/time_series/rule'
 require 'redis/time_series/info'
 require 'redis/time_series/sample'
